pyproject.toml
[tool.poetry]
name = "ad-data-pipeline"
version = "0.1.0"
description = "Test task pipeline for parsing and analyzing Facebook Ad Library data"
authors = ["Artur Chukhrai"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2"
numpy = "<2.0.0"
pyspark = "=3.5.6"
requests = "^2.32"
langdetect = "^1.0"
pydantic = "^2.6"
python-dotenv = "^1.0"
cryptography = "^42.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-cov = "^5.0"
black = "^24.3"
ruff = "^0.8"
mypy = "^1.13"
pre-commit = "^4.0"
types-requests = "^2.31"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
exclude_lines = ["if __name__ == .__main__.:"]

# Ruff configuration
[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# Black configuration
[tool.black]
target-version = ['py310']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "pyspark.*",
    "langdetect.*",
]
ignore_missing_imports = true
